pyproject.toml
[project]
name = "drive-thru-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "pyaudio>=0.2.14",
    "pyttsx3>=2.99",
    "speechrecognition>=3.14.3",
]
